refactor: Move files to details package for better organization
- Moved files related to movie details use case to the `presentation.details package to improve code structure and organization
