+++ presentation/src/test/java/com/curtesmalteser/presentation/SetupMovieDetailsUseCaseTest.kt
@@ -3,6 +3,7 @@ package com.curtesmalteser.presentation
 import com.curtesmalteser.popularmovies.data.MovieData
 import com.curtesmalteser.popularmovies.data.MoviesModelData
 import com.curtesmalteser.popularmovies.repository.IMoviesRepository
+import com.curtesmalteser.presentation.details.SetupMovieDetailsUseCase
 import com.curtesmalteser.presentation.mock.repository.MovieDetailsRepositoryMock
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.shouldBe
